+++ advanced-responsive-video-embedder.php
@@ -16,7 +16,7 @@
  * Plugin Name:       ARVE Advanced Responsive Video Embedder
  * Plugin URI:        https://nextgenthemes.com/plugins/advanced-responsive-video-embedder-pro/
  * Description:       Easy responsive video embeds via URL (like WordPress) or Shortcodes. Supports almost anything you can imagine.
- * Version:           7.0.4
+ * Version:           7.0.5
  * Author:            Nicolas Jonas
  * Author URI:        https://nextgenthemes.com
  * License:           GPL-3.0
@@ -31,6 +31,8 @@
 if ( ! defined( 'WPINC' ) ) {
 	die;
 }
+
+define( 'ARVE_PRO_VERSION_REQUIRED', '2.5.4' );
 
 /**
  * The code that runs during plugin activation.